move generateWehookPath to path.go
