+++ internal/webhook/external/path.go
@@ -1,0 +1,13 @@
+package external
+
+import (
+	"crypto/sha256"
+	"fmt"
+)
+
+func GenerateWebhookPath(project, provider, token string) string {
+	input := []byte(project + provider + token)
+	h := sha256.New()
+	h.Write(input)
+	return fmt.Sprintf("/%x", h.Sum(nil))
+}
